+++ modules/shared/core/src/sap/commerce/toolset/settings/state/DeveloperSettingsState.kt
@@ -28,7 +28,6 @@ data class DeveloperSettingsState(
     @JvmField @OptionTag val beanSystemSettings: BeanSystemSettingsState = BeanSystemSettingsState(),
     @JvmField @OptionTag val typeSystemSettings: TypeSystemSettingsState = TypeSystemSettingsState(),
     @JvmField @OptionTag val cngSettings: CngSettingsState = CngSettingsState(),
-    @JvmField @OptionTag val bpSettings: BpSettingsState = BpSettingsState(),
     @JvmField @OptionTag val flexibleSearchSettings: FlexibleSearchSettingsState = FlexibleSearchSettingsState(),
     @JvmField @OptionTag val polyglotQuerySettings: PolyglotQuerySettingsState = PolyglotQuerySettingsState(),
     @JvmField @OptionTag val impexSettings: ImpExSettingsState = ImpExSettingsState(),