+++ CHANGELOG.md
@@ -37,6 +37,9 @@
 ### `FlexibleSearch` enhancements
 - Use IntelliJ extension point for Code Folding settings [#1549](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1549)
 
+### `Business Process` enhancements
+- Use IntelliJ extension point for Code Folding settings [#1550](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1550)
+
 ### `Other` enhancements
 - Migrated to immutable Settings [#1527](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1527)
 - Added possibility to specify `timeout` per connection settings [#1540](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1540)